requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-environ>=0.11.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
django-allauth>=0.57.0
django-filter>=23.5
django-extensions>=3.2.0
whitenoise>=6.6.0
gunicorn>=21.2.0
python-decouple>=3.8
django-ratelimit>=4.1.0
django-audit-log>=0.7.0
django-notifications-hq>=1.8.0
twilio>=8.10.0
sendgrid>=6.10.0
cryptography>=41.0.0
django-otp>=1.2.0
qrcode>=7.4.0
